Separate config and options specific code for node from start
First refactoring towards separating out the code that launches
networking from the code that launches the kubelet. Ensures that node
options (the part that converts nodeconfig to kubelet config) is
completely separate from how processes are started.
