+++ pkg/cmd/server/kubernetes/node/node_config.go
@@ -4,9 +4,7 @@ import (
 	"crypto/tls"
 	"fmt"
 	"net"
-	"strconv"
 	"strings"
-	"time"
 
 	"github.com/golang/glog"
 
@@ -15,11 +13,9 @@ import (
 	kerrs "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
-	kerrors "k8s.io/apimachinery/pkg/util/errors"
 	clientgoclientset "k8s.io/client-go/kubernetes"
 	kv1core "k8s.io/client-go/kubernetes/typed/core/v1"
 	"k8s.io/client-go/util/cert"
-	kubeproxyoptions "k8s.io/kubernetes/cmd/kube-proxy/app"
 	kubeletapp "k8s.io/kubernetes/cmd/kubelet/app"
 	kubeletoptions "k8s.io/kubernetes/cmd/kubelet/app/options"
 	"k8s.io/kubernetes/pkg/apis/componentconfig"
@@ -30,16 +26,12 @@ import (
 	"k8s.io/kubernetes/pkg/cloudprovider"
 	"k8s.io/kubernetes/pkg/kubelet"
 	dockertools "k8s.io/kubernetes/pkg/kubelet/dockershim/libdocker"
-	kubeletcni "k8s.io/kubernetes/pkg/kubelet/network/cni"
 	kubeletserver "k8s.io/kubernetes/pkg/kubelet/server"
-	kubelettypes "k8s.io/kubernetes/pkg/kubelet/types"
 
 	osclient "github.com/openshift/origin/pkg/client"
 	configapi "github.com/openshift/origin/pkg/cmd/server/api"
 	"github.com/openshift/origin/pkg/cmd/server/crypto"
 	cmdutil "github.com/openshift/origin/pkg/cmd/util"
-	cmdflags "github.com/openshift/origin/pkg/cmd/util/flags"
-	"github.com/openshift/origin/pkg/cmd/util/variable"
 	"github.com/openshift/origin/pkg/dns"
 	"github.com/openshift/origin/pkg/network"
 	networkapi "github.com/openshift/origin/pkg/network/apis/network"
@@ -85,7 +77,16 @@ type NodeConfig struct {
 	SDNProxy network.ProxyInterface
 }
 
-func BuildKubernetesNodeConfig(options configapi.NodeConfig, enableProxy, enableDNS bool) (*NodeConfig, error) {
+func New(options configapi.NodeConfig, server *kubeletoptions.KubeletServer, proxyconfig *componentconfig.KubeProxyConfiguration, enableProxy, enableDNS bool) (*NodeConfig, error) {
+	if options.NodeName == "localhost" {
+		glog.Warningf(`Using "localhost" as node name will not resolve from all locations`)
+	}
+
+	clientCAs, err := cert.NewPool(options.ServingInfo.ClientCA)
+	if err != nil {
+		return nil, err
+	}
+
 	originClient, _, err := configapi.GetOpenShiftClient(options.MasterKubeConfig, options.MasterClientConnectionOverrides)
 	if err != nil {
 		return nil, err
@@ -108,126 +109,7 @@ func BuildKubernetesNodeConfig(options configapi.NodeConfig, enableProxy, enable
 		return nil, err
 	}
 
-	if options.NodeName == "localhost" {
-		glog.Warningf(`Using "localhost" as node name will not resolve from all locations`)
-	}
-
-	clientCAs, err := cert.NewPool(options.ServingInfo.ClientCA)
-	if err != nil {
-		return nil, err
-	}
-
-	imageTemplate := variable.NewDefaultImageTemplate()
-	imageTemplate.Format = options.ImageConfig.Format
-	imageTemplate.Latest = options.ImageConfig.Latest
-
-	var path string
-	var fileCheckInterval int64
-	if options.PodManifestConfig != nil {
-		path = options.PodManifestConfig.Path
-		fileCheckInterval = options.PodManifestConfig.FileCheckIntervalSeconds
-	}
-
-	kubeAddressStr, kubePortStr, err := net.SplitHostPort(options.ServingInfo.BindAddress)
-	if err != nil {
-		return nil, fmt.Errorf("cannot parse node address: %v", err)
-	}
-	kubePort, err := strconv.Atoi(kubePortStr)
-	if err != nil {
-		return nil, fmt.Errorf("cannot parse node port: %v", err)
-	}
-
 	if err = validateNetworkPluginName(originClient, options.NetworkConfig.NetworkPluginName); err != nil {
-		return nil, err
-	}
-
-	// Defaults are tested in TestKubeletDefaults
-	server := kubeletoptions.NewKubeletServer()
-	// Adjust defaults
-	server.RequireKubeConfig = true
-	server.KubeConfig.Default(options.MasterKubeConfig)
-	server.PodManifestPath = path
-	server.RootDirectory = options.VolumeDirectory
-	server.NodeIP = options.NodeIP
-	server.HostnameOverride = options.NodeName
-	server.AllowPrivileged = true
-	server.RegisterNode = true
-	server.Address = kubeAddressStr
-	server.Port = int32(kubePort)
-	server.ReadOnlyPort = 0        // no read only access
-	server.CAdvisorPort = 0        // no unsecured cadvisor access
-	server.HealthzPort = 0         // no unsecured healthz access
-	server.HealthzBindAddress = "" // no unsecured healthz access
-	server.ClusterDNS = []string{options.DNSIP}
-	server.ClusterDomain = options.DNSDomain
-	server.NetworkPluginName = options.NetworkConfig.NetworkPluginName
-	server.HostNetworkSources = []string{kubelettypes.ApiserverSource, kubelettypes.FileSource}
-	server.HostPIDSources = []string{kubelettypes.ApiserverSource, kubelettypes.FileSource}
-	server.HostIPCSources = []string{kubelettypes.ApiserverSource, kubelettypes.FileSource}
-	server.HTTPCheckFrequency = metav1.Duration{Duration: time.Duration(0)} // no remote HTTP pod creation access
-	server.FileCheckFrequency = metav1.Duration{Duration: time.Duration(fileCheckInterval) * time.Second}
-	server.KubeletFlags.ContainerRuntimeOptions.PodSandboxImage = imageTemplate.ExpandOrDie("pod")
-	server.LowDiskSpaceThresholdMB = 256 // this the previous default
-	server.CPUCFSQuota = true            // enable cpu cfs quota enforcement by default
-	server.MaxPods = 250
-	server.PodsPerCore = 10
-	server.CgroupDriver = "systemd"
-	server.DockerExecHandlerName = string(options.DockerConfig.ExecHandlerName)
-	server.RemoteRuntimeEndpoint = options.DockerConfig.DockerShimSocket
-	server.RemoteImageEndpoint = options.DockerConfig.DockerShimSocket
-	server.DockershimRootDirectory = options.DockerConfig.DockershimRootDirectory
-
-	if network.IsOpenShiftNetworkPlugin(server.NetworkPluginName) {
-		// set defaults for openshift-sdn
-		server.HairpinMode = componentconfig.HairpinNone
-	}
-
-	// prevents kube from generating certs
-	server.TLSCertFile = options.ServingInfo.ServerCert.CertFile
-	server.TLSPrivateKeyFile = options.ServingInfo.ServerCert.KeyFile
-
-	containerized := cmdutil.Env("OPENSHIFT_CONTAINERIZED", "") == "true"
-	server.Containerized = containerized
-
-	// force the authentication and authorization
-	// Setup auth
-	authnTTL, err := time.ParseDuration(options.AuthConfig.AuthenticationCacheTTL)
-	if err != nil {
-		return nil, err
-	}
-	server.Authentication = componentconfig.KubeletAuthentication{
-		X509: componentconfig.KubeletX509Authentication{
-			ClientCAFile: options.ServingInfo.ClientCA,
-		},
-		Webhook: componentconfig.KubeletWebhookAuthentication{
-			Enabled:  true,
-			CacheTTL: metav1.Duration{Duration: authnTTL},
-		},
-		Anonymous: componentconfig.KubeletAnonymousAuthentication{
-			Enabled: true,
-		},
-	}
-	authzTTL, err := time.ParseDuration(options.AuthConfig.AuthorizationCacheTTL)
-	if err != nil {
-		return nil, err
-	}
-	server.Authorization = componentconfig.KubeletAuthorization{
-		Mode: componentconfig.KubeletAuthorizationModeWebhook,
-		Webhook: componentconfig.KubeletWebhookAuthorization{
-			CacheAuthorizedTTL:   metav1.Duration{Duration: authzTTL},
-			CacheUnauthorizedTTL: metav1.Duration{Duration: authzTTL},
-		},
-	}
-
-	// resolve extended arguments
-	// TODO: this should be done in config validation (along with the above) so we can provide
-	// proper errors
-	if err := cmdflags.Resolve(options.KubeletArguments, server.AddFlags); len(err) > 0 {
-		return nil, kerrors.NewAggregate(err)
-	}
-
-	proxyconfig, err := buildKubeProxyConfig(options)
-	if err != nil {
 		return nil, err
 	}
 
@@ -241,13 +123,6 @@ func BuildKubernetesNodeConfig(options configapi.NodeConfig, enableProxy, enable
 		if err != nil {
 			return nil, fmt.Errorf("SDN initialization failed: %v", err)
 		}
-
-		// SDN plugin pod setup/teardown is implemented as a CNI plugin
-		server.NetworkPluginName = kubeletcni.CNIPluginName
-		server.NetworkPluginDir = kubeletcni.DefaultNetDir
-		server.CNIConfDir = kubeletcni.DefaultNetDir
-		server.CNIBinDir = kubeletcni.DefaultCNIDir
-		server.HairpinMode = componentconfig.HairpinNone
 	}
 
 	deps, err := kubeletapp.UnsecuredKubeletDeps(server)
@@ -298,7 +173,7 @@ func BuildKubernetesNodeConfig(options configapi.NodeConfig, enableProxy, enable
 		BindAddress: options.ServingInfo.BindAddress,
 
 		AllowDisabledDocker: options.AllowDisabledDocker,
-		Containerized:       containerized,
+		Containerized:       server.Containerized,
 
 		Client:                kubeClient,
 		ExternalKubeClientset: externalKubeClient,
@@ -368,69 +243,6 @@ func BuildKubernetesNodeConfig(options configapi.NodeConfig, enableProxy, enable
 	}
 
 	return config, nil
-}
-
-func buildKubeProxyConfig(options configapi.NodeConfig) (*componentconfig.KubeProxyConfiguration, error) {
-	proxyOptions, err := kubeproxyoptions.NewOptions()
-	if err != nil {
-		return nil, err
-	}
-	// get default config
-	proxyconfig := proxyOptions.GetConfig()
-
-	// BindAddress - Override default bind address from our config
-	addr := options.ServingInfo.BindAddress
-	host, _, err := net.SplitHostPort(addr)
-	if err != nil {
-		return nil, fmt.Errorf("The provided value to bind to must be an ip:port %q", addr)
-	}
-	ip := net.ParseIP(host)
-	if ip == nil {
-		return nil, fmt.Errorf("The provided value to bind to must be an ip:port: %q", addr)
-	}
-	proxyconfig.BindAddress = ip.String()
-
-	// HealthzPort, HealthzBindAddress - disable
-	proxyconfig.HealthzBindAddress = ""
-	proxyconfig.MetricsBindAddress = ""
-
-	// OOMScoreAdj, ResourceContainer - clear, we don't run in a container
-	oomScoreAdj := int32(0)
-	proxyconfig.OOMScoreAdj = &oomScoreAdj
-	proxyconfig.ResourceContainer = ""
-
-	// use the same client as the node
-	proxyconfig.ClientConnection.KubeConfigFile = options.MasterKubeConfig
-
-	// ProxyMode, set to iptables
-	proxyconfig.Mode = "iptables"
-
-	// IptablesSyncPeriod, set to our config value
-	syncPeriod, err := time.ParseDuration(options.IPTablesSyncPeriod)
-	if err != nil {
-		return nil, fmt.Errorf("Cannot parse the provided ip-tables sync period (%s) : %v", options.IPTablesSyncPeriod, err)
-	}
-	proxyconfig.IPTables.SyncPeriod = metav1.Duration{
-		Duration: syncPeriod,
-	}
-	masqueradeBit := int32(0)
-	proxyconfig.IPTables.MasqueradeBit = &masqueradeBit
-
-	// PortRange, use default
-	// HostnameOverride, use default
-	// ConfigSyncPeriod, use default
-	// MasqueradeAll, use default
-	// CleanupAndExit, use default
-	// KubeAPIQPS, use default, doesn't apply until we build a separate client
-	// KubeAPIBurst, use default, doesn't apply until we build a separate client
-	// UDPIdleTimeout, use default
-
-	// Resolve cmd flags to add any user overrides
-	if err := cmdflags.Resolve(options.ProxyArguments, proxyOptions.AddFlags); len(err) > 0 {
-		return nil, kerrors.NewAggregate(err)
-	}
-
-	return proxyconfig, nil
 }
 
 func validateNetworkPluginName(originClient *osclient.Client, pluginName string) error {
